Multi-site on the same jboss app server.
We decided to treat this on database, I've created a master entity called Site to keep all the domains of the server.

In contentServlet I retrieve the domain being accessed and pass this to business layer, the main entities will have a foreign key called site_id to filter the content.

To make the many domains work, we need to configure them on jbossweb.sar/server.xml, like this:

<Host name="ahycms.org" autoDeploy="true"
      deployOnStartup="true"
      deployXML="false">
      <Alias>domain1.com</Alias>
      <Alias>domain2.com</Alias>
</Host>